requirements.txt
anyio==3.6.2
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
cloudinary==1.32.0
cryptography==40.0.2
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post1
et-xmlfile==1.1.0
fastapi==0.95.1
h11==0.14.0
httptools==0.5.0
idna==3.4
mysqlclient==2.1.1
openpyxl==3.1.2
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.7
pypng==0.20220715.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
qrcode==7.4.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.9
starlette==0.26.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.1
